requirements.txt
numpy
scipy
astropy
healpy
fitsio
numba
setuptools
gitpython
mpi4py
